requirements.txt
alabaster==0.7.16
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Cython==3.0.2
dash==3.1.1
docutils==0.20.1
Flask==3.1.1
fonttools==4.61.1
furo==2023.9.10
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.1
numpy==2.4.2
packaging==25.0
pandas==2.2.2
pillow==12.1.0
plotly==6.2.0
pluggy==1.6.0
psutil==5.9.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==7.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.4
retrying==1.4.0
ruff==0.0.291
setuptools==73.0.1
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==7.2.6
sphinx-basic-ng==1.0.0b2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.44.0
zipp==3.23.0
